// src/app/shared/models/employee-dto.model.ts
export interface EmployeeDto {
  employeeId: number;
  employeeName: string;
  empShort: string;
  empPerson?: string;
  empPosition?: string;
  empEmail?: string;
  empMobile?: string;
  empAddress?: string;
  empType?: string;
  bitIsActive?: boolean;
  dateCreated?: string;
  bitIsDelete?: boolean;
  empPassword?: string;
  bintUserId?: number;
  empStartDate?: string;
}