// src/app/shared/models/employee.model.ts
export interface Employee {
  employeeId: number;
  employeeName: string;
  empShort: string;
  empPerson: string | null;
  empPosition: string | null;
  empEmail: string | null;
  empMobile: string | null;
  empAddress: string | null;
  empType: string | null;
  bitIsActive: boolean | null;
  dateCreated: string | null;
  bitIsDelete: boolean | null;
  empPassword: string | null;
  bintUserId: number | null;
  empStartDate: string | null;
}